pyproject.toml
[project]
name = "onboarding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "httpx>=0.26.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.40",
    "openai>=1.10.0",
    "pydantic>=2.5.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=16.0",
]
